The return type of the Area property of ConnectedComponent was changed in the Native library.

+++ src/Magick.NET/Shared/Types/ConnectedComponent.cs
@@ -23,7 +23,7 @@ namespace ImageMagick
     {
         private ConnectedComponent(IntPtr instance)
         {
-            Area = (int)NativeConnectedComponent.GetArea(instance);
+            Area = NativeConnectedComponent.GetArea(instance);
             Centroid = PointD.FromPointInfo(NativeConnectedComponent.GetCentroid(instance));
             Color = NativeConnectedComponent.GetColor(instance);
             Height = NativeConnectedComponent.GetHeight(instance);
